+++ tinymce/tests/test_widgets.py
@@ -1,11 +1,25 @@
 # coding: utf-8
 
+from contextlib import contextmanager
+
 from django import forms
 from django.test import TestCase
 from django.test.utils import override_settings
 from django.utils.translation import override
 
+import tinymce.settings
 from tinymce.widgets import get_language_config, TinyMCE
+
+
+@contextmanager
+def override_tinymce_settings(settings_dict):
+    saved_values = {}
+    for setting, value in settings_dict.items():
+        saved_values[setting] = getattr(tinymce.settings, setting)
+        setattr(tinymce.settings, setting, value)
+    yield
+    for setting in settings_dict.keys():
+        setattr(tinymce.settings, setting, saved_values[setting])
 
 
 @override_settings(LANGUAGES=[('en', 'English')])
@@ -59,6 +73,28 @@ class TestWidgets(TestCase):
             'foobar', 'lorem ipsum', attrs={'id': 'id_foobar', 'class': 'foo'}
         )
         self.assertIn('class="foo tinymce"', html)
+
+    @override_settings(TINYMCE_INCLUDE_JQUERY=False)
+    def test_tinymce_widget_media(self):
+        widget = TinyMCE()
+        self.assertEqual(
+            widget.media.render_js(),
+            [
+                '<script type="text/javascript" src="/tinymce/compressor/"></script>',
+                '<script type="text/javascript" src="/static/django_tinymce/jquery-1.9.1.min.js"></script>',
+                '<script type="text/javascript" src="/static/django_tinymce/init_tinymce.js"></script>',
+            ]
+        )
+        self.assertEqual(list(widget.media.render_css()), [])
+        with override_tinymce_settings({'USE_COMPRESSOR': False, 'INCLUDE_JQUERY': False}):
+            widget = TinyMCE()
+            self.assertEqual(
+                widget.media.render_js(),
+                [
+                    '<script type="text/javascript" src="/static/tiny_mce/tiny_mce.js"></script>',
+                    '<script type="text/javascript" src="/static/django_tinymce/init_tinymce.js"></script>',
+                ]
+            )
 
     def test_tinymce_widget_required(self):
         """